feat: add a config file to simple edit the default transaction and block.
