+++ README.md
@@ -32,9 +32,12 @@ cargo run
 ```
 For now, the flags are not working, they are just here for development annotation.
 
-### Basic usage
+## Basic usage
 
 Type `help` on the CLI to see the available commands.
 ```b
 > help
 ```
+
+## Advaced usage 
+In```/src/reference_implementation/``` you will find: ```defaults.json```, ```api.rs``` , ```cli.rs```,  ```read_defaults.rs```, you can change the json file to personalize the defaults to random generation of blocks or transactions, or you can read this folder to learn how to use by importing our lib, you can use Misfit_core directly on your project generating direclty your use cases.
