+++ src/lib.rs
@@ -1,3 +1,3 @@
 pub mod transaction;
-pub mod regtest_pack;
+pub mod blockchain_manager;
 pub mod block;
